split prod/dev webpack builds and define constants including version
