+++ .gitignore
@@ -3,4 +3,3 @@
 build/
 dist/
 node_modules/
-betterdgg/modules/version.js
