+++ betterdgg/modules/version.js
@@ -1,0 +1,2 @@
+// set by webpack
+export default VERSION
